32-bit types diff

+++ lib/MyersDiff/MyersDiff.cpp
@@ -2,19 +2,19 @@
 
 namespace Myers {
 
-Snake::Snake(size_t begin_x, size_t begin_y, size_t width)
+Snake::Snake(uint32_t begin_x, uint32_t begin_y, uint32_t width)
     : begin_x_(begin_x), begin_y_(begin_y), width_(width) {
 }
 
-std::pair<size_t, size_t> Snake::Begin() {
+std::pair<uint32_t, uint32_t> Snake::Begin() {
     return {begin_x_, begin_y_};
 }
 
-size_t Snake::Width() {
+uint32_t Snake::Width() {
     return width_;
 }
 
-std::pair<size_t, size_t> Snake::End() {
+std::pair<uint32_t, uint32_t> Snake::End() {
     return {begin_x_ + width_, begin_y_ + width_};
 }
